requirements.txt
# python -m pip install --upgrade pip setuptools wheel


# opencv-python-headless 
pytesseract 
pillow
streamlit
